capstone-website-new-edition/requirements.txt
# Flask and web requirements
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
python-dotenv==0.19.0

# Database
psycopg2-binary==2.9.1

# Forms
Flask-WTF==0.15.1
WTForms==2.3.3
email-validator==1.1.3

# Security
Flask-Bcrypt==0.7.1
bcrypt==3.2.0

# SCSS compilation requirements
libsass==0.22.0
watchdog==3.0.0
Flask-Assets==2.1.0  # Optional for Flask integration
